chasing compiler warnings
